init: load most of the menu from ROM and zero BSS
The BIOS only loads the first 0x1000 bytes on our first run no matter
what the header says. Load again over the top to make sure we get the
lot. Means the moving parts have to be in the first 0x1000 bytes - and
this invokes an intrinsic shift operator from libgcc. Yuck.