Add stubloader
This allows us to transfer power over the BIOS boot barrier.

Normally when you boot a disc (emulated or no) the BIOS basically
wipes all of RAM. You could trick over it with the UBC but only by
breaking on BIOS addresses, which is a grubby hack. So this is a
different hack.

The stubloader is a tiny bootable iso - just an IP/AIP, less than
the 0x8000 byte limit so that no further loading is required.
This does two things:
- region frees the system (passively just by being all-region), and
- installs a return-to-menu hook into some of the BIOS' vector table
in HiRAM.
After this it then flips to emulating "out.desc" and manually loads
and boots the disc header.

The BIOS vector table location is chosen somewhat arbitrarily as
somewhere without any default vectors. It may be that some games or
peripherals do use it which will be a problem, but moving it around
shouldn't really be an issue. If worst comes to worst it's a small
enough amount of code you could probably hide it on the stack...
and the menu can opt out of using it for specific games and boot them
"natively" anyway. That'd include patching just the first sector or w/e
using the descriptor mechanism.

The boot process includes loading a file to the "1st READ ADDRESS",
normally during logo display. We do it afterward (in darkness) and
so this may add a couple of seconds to game boot. The Right Way to
do this would be to patch the stubloader header to do the load
during logo time (as well as set master and slave stack pointers
directly). As the stubloader currently relocates itself into
LoRAM unconditionally, this might mean having two stub-halves linked
to low and high work RAM, respectively, so that the appropriate one
can be chosen.

This is still a bit incomplete. It doesn't set stack pointers
(master or slave) if the disc image requests it. It also doesn't
wipe the work RAM clean before booting. As I learned - when
the menu would crash after returning from a game - some code
implicitly depends on this zeroing and it must be implemented. This
means hiding a little wiper on the stack to clean up afterward.