Implement new CD descriptor format
New format uses segments - a part of the disc with uniform
track:index:q_mode:file - rather than tracks. This makes the reader code
simpler as well as more flexible (eg. can patch just the start of a
track).

- split satisfier-types.h out of satisfier.h; the former is to be
identical to the satisfier.h in firmware common/ as well as yabause
- implement cue2desc, with a proper multifile parser
- implement image2desc, a commandline tool