multidisc booting
Multi-disc titles are identified by being in a folder on their own, with the
last digit in the filename indicating the disc number. The rest of the filename
has to match exactly. For example:

    Rampo (Disc 1).cue
    Rampo (Disc 2).cue

.bin and so on are ignored for this process, but if you have .cue files that
reference .iso files you should rename them (and fix the name in the .cue).

These are then turned into out1.desc, out2.desc et cetera, which is how the
firmware identifies that they're to be treated as multidisc - the digit right
before the dot.

Up to 9 discs are supported.