smpc: make sure all calls are made before iapetus per_init()
the iapetus SMPC handler assumes that all interrupts are for peripheral
inputs, and it will happily overflow the per[] array with other inputs.

this would cause the text layers not to initialise, because the
vdp2_settings struct was clobbered.

fix by doing the couple of simple SMPC jobs first, then handing over to
iapetus for all the input work.